% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizR.R
\name{add_question}
\alias{add_question}
\title{Add question to group}
\usage{
add_question(group, question)
}
\arguments{
\item{group}{Group object}

\item{question}{Question to add}
}
\description{
Add question to group
}

