% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizR.R
\name{addQuestion}
\alias{addQuestion}
\title{Add question to group}
\usage{
addQuestion(group, question)
}
\arguments{
\item{group}{Group object}

\item{question}{Question to add}
}
\description{
Add question to group
}

