% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizR.R
\name{cloze_coefficients}
\alias{cloze_coefficients}
\title{Return coefficients of cloze fields over their sum}
\usage{
cloze_coefficients(question)
}
\arguments{
\item{question}{A question}
}
\description{
Return coefficients of cloze fields over their sum
}

