% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloze.R
\name{cloze_field_points}
\alias{cloze_field_points}
\title{Return list of cloze fields points}
\usage{
cloze_field_points(question)
}
\arguments{
\item{question}{A question}
}
\description{
Return list of cloze fields points
}
