% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizR.R
\name{compute_results}
\alias{compute_results}
\title{Compute full results of given quiz}
\usage{
compute_results(quiz, filename = NULL)
}
\arguments{
\item{quiz}{Quiz}

\item{filename}{CSV file of results}
}
\description{
Compute full results of given quiz
}

