% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizR.R
\name{correct_question}
\alias{correct_question}
\title{Correct question}
\usage{
correct_question(question, env, guess)
}
\arguments{
\item{question}{A question}

\item{env}{The environment in which to evaluate answers}

\item{guess}{A guess}
}
\description{
Correct question
}

