% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizR.R
\name{distinct_data}
\alias{distinct_data}
\title{Check that data do not overlap}
\usage{
distinct_data(quiz, lang)
}
\arguments{
\item{quiz}{A quiz object}

\item{lang}{Optional additionnal language object}
}
\value{
Return TRUE if data do not overlap
}
\description{
Check that data do not overlap
}
