% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizR.R
\name{distinct_language}
\alias{distinct_language}
\title{Check if languages are commutative chunks of code}
\usage{
distinct_language(lang1, lang2)
}
\arguments{
\item{lang1}{First chunk of code}

\item{lang2}{Second chunk of code}
}
\value{
Return TRUE if the two chunks of code are commutative
}
\description{
Check if languages are commutative chunks of code
}

