% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizR.R
\name{evalAnswers}
\alias{evalAnswers}
\title{Evaluate list of answers in environment}
\usage{
evalAnswers(answers, env)
}
\arguments{
\item{answers}{List of answers}

\item{env}{Environment in which to evaluate}
}
\description{
Evaluate list of answers in environment
}

