% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expr}
\alias{expr}
\title{Helper function to specify list of languages}
\usage{
expr(...)
}
\arguments{
\item{Expressions}{}
}
\value{
A quoted expression or a list of quoted expression
}
\description{
Helper function
}

