% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_data_file}
\alias{generate_data_file}
\title{Generate data file}
\usage{
generate_data_file(quiz, data.name = paste0(quiz$title, "-data.R"),
  language = NULL)
}
\arguments{
\item{quiz}{Quiz}

\item{data.name}{Name of the data file}

\item{language}{Additionnal language for validation}
}
\description{
Generate data file
}
