% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_files}
\alias{generate_files}
\title{Generate XML Moodle quiz file and data file}
\usage{
generate_files(quiz, data.name = paste0(quiz$title, "-data.R"),
  quiz.name = paste0(quiz$title, "-quiz.xml"), language)
}
\arguments{
\item{quiz}{Quiz}

\item{data.name}{Name of data file}

\item{quiz.name}{Name of XML file}

\item{language}{Additionnal language for validation}
}
\description{
Generate XML Moodle quiz file and data file
}

