% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_XML}
\alias{generate_XML}
\title{Generate XML Moodle quiz file}
\usage{
generate_XML(quiz, quiz.name = paste0(quiz$title, "-quiz.xml"),
  language = NULL, feedback = TRUE, eval = feedback, answer = feedback)
}
\arguments{
\item{quiz}{Quiz}

\item{quiz.name}{Name of generated file}

\item{language}{Additionnal language for validation}

\item{feedback}{Whether to add feedback in XML}

\item{eval}{Whether to evaluate answers in feedback}
}
\description{
Generate XML Moodle quiz file
}
