% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generateFiles}
\alias{generateFiles}
\title{Generate XML Moodle quiz file and data file}
\usage{
generateFiles(quiz, data.name = paste0(quiz$title, "-data"),
  quiz.name = paste0(quiz$title, "-quiz"))
}
\arguments{
\item{quiz}{Quiz}

\item{data.name}{Name of data file}

\item{quiz.name}{Name of XML file}
}
\description{
Generate XML Moodle quiz file and data file
}

