% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloze.R
\name{get_cloze_num}
\alias{get_cloze_num}
\title{Number of fields in cloze question}
\usage{
get_cloze_num(text)
}
\arguments{
\item{text}{Some character string}
}
\description{
Number of fields in cloze question
}
