% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizR.R
\name{get_num}
\alias{get_num}
\title{Number of real questions (except description questions)}
\usage{
get_num(group)
}
\arguments{
\item{group}{Group object}
}
\description{
Number of real questions (except description questions)
}

