% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizR.R
\name{getNum}
\alias{getNum}
\title{Number of real questions (except description questions)}
\usage{
getNum(group)
}
\arguments{
\item{group}{Group object}
}
\description{
Number of real questions (except description questions)
}

