% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizR.R
\name{Group}
\alias{Group}
\title{Create a group object}
\usage{
Group(title, type, num, data, hidden.data, questions)
}
\arguments{
\item{title}{Title of group}

\item{type}{Type of group}

\item{num}{Number of questions of group if of random type}
}
\description{
Create a group object
}

