% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizR.R
\name{match_answers}
\alias{match_answers}
\title{Match guess with list of evaluated answers}
\usage{
match_answers(eval_answers, guess, dist, epsilon)
}
\arguments{
\item{eval_answers}{List of evaluated answers}

\item{guess}{A guess}

\item{dist}{Maximum edit distance for a match}

\item{epsilon}{Relative error for numeric answers}
}
\description{
Match guess with list of evaluated answers
}
