% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizR.R
\name{matchAnswers}
\alias{matchAnswers}
\title{Match guess with list of evaluated answers}
\usage{
matchAnswers(evalAnswers, guess, dist, epsilon)
}
\arguments{
\item{evalAnswers}{List of evaluated answers}

\item{guess}{A guess}

\item{dist}{Maximum edit distance for a match}

\item{epsilon}{Relative error for numeric answers}
}
\description{
Match guess with list of evaluated answers
}

