% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizR.R
\docType{data}
\name{question_types}
\alias{question_types}
\title{Character vector of all handled question types}
\format{An object of class \code{character} of length 4.}
\usage{
question_types
}
\description{
Character vector of all handled question types
}
\keyword{datasets}
