% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizR.R
\name{Quiz}
\alias{Quiz}
\title{Create a quiz object}
\usage{
Quiz(title, groups, data, hidden.data)
}
\arguments{
\item{title}{Title of the quiz}

\item{groups}{List of groups}
}
\value{
A quiz object
}
\description{
Create a quiz object
}

