% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizR.R
\name{replace_answers}
\alias{replace_answers}
\title{Replace data in list of answers}
\usage{
replace_answers(answers, data)
}
\arguments{
\item{answers}{List of answers}

\item{data}{Data to replace}
}
\description{
Replace data in list of answers
}

