% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizR.R
\name{replace_language}
\alias{replace_language}
\title{Replace some symbols in language object by aliases}
\usage{
replace_language(lang, aliases)
}
\description{
Replace some symbols in language object by aliases
}
