% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizR.R
\name{replace_text}
\alias{replace_text}
\title{Replace hidden data in text of all questions}
\usage{
replace_text(obj, lang = NULL)
}
\arguments{
\item{obj}{A quiz object}

\item{lang}{An optional language object}
}
\value{
A character string with some of the variables replaced
}
\description{
Replace hidden data in text of all questions
}
