% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizR.R
\name{replace_text_data}
\alias{replace_text_data}
\title{Replace symbols of R chunks of code in text}
\usage{
replace_text_data(text, data)
}
\arguments{
\item{text}{A character string written in R markdown}

\item{data}{A language object defining variables to be replaced}
}
\value{
A character string with some of the variables replaced
}
\description{
Replace symbols of R chunks of code in text
}
