% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizR.R
\name{replace_text_env}
\alias{replace_text_env}
\title{Replace symbols of R chunks of code in text}
\usage{
replace_text_env(text, env)
}
\arguments{
\item{text}{A character string written in R markdown}

\item{env}{An environment containing variables to be replaced}
}
\value{
A character string with some of the variables replaced
}
\description{
Replace symbols of R chunks of code in text
}
