% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_question.R
\docType{data}
\name{SimpleQuestion}
\alias{SimpleQuestion}
\title{Base class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SimpleQuestion
}
\description{
Base class
}
\keyword{datasets}
