% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizR.R
\name{validate_quiz}
\alias{validate_quiz}
\title{Validate a quiz object}
\usage{
validate_quiz(quiz, lang)
}
\arguments{
\item{quiz}{A quiz object}

\item{lang}{Additionnal language object}
}
\value{
Return TRUE if quiz is a valid one
}
\description{
Validate a quiz object
}
