% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizR.R
\name{write_grades}
\alias{write_grades}
\title{Write grades from answers in csv file}
\usage{
write_grades(quiz, input = NULL, output = NULL, lang = NULL)
}
\arguments{
\item{quiz}{Quiz}

\item{input}{CSV file of results}

\item{output}{CSV file of grades}

\item{lang}{Chunk of code to pass validation}
}
\description{
Write grades from answers in csv file
}

