% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizR.R
\name{compute_grades}
\alias{compute_grades}
\title{Compute grades for given quiz and results}
\usage{
compute_grades(quiz, filename = NULL)
}
\arguments{
\item{quiz}{Quiz}

\item{filename}{CSV file of results}
}
\description{
Compute grades for given quiz and results
}

