% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizR.R
\name{Group}
\alias{Group}
\title{Create a group object}
\usage{
Group(title, type, num, data, hidden.data, hidden.seed = NULL, questions)
}
\arguments{
\item{title}{Title of group}

\item{type}{Type of group}

\item{num}{Number of questions of group if of random type}

\item{data}{Language object defining the data}

\item{hidden.data}{Language object defining the hidden data}

\item{hidden.seed}{The seed used to generate hidden data}

\item{questions}{List of questions}
}
\description{
Create a group object
}
