% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizR.R
\name{Question}
\alias{Question}
\title{Create a question object}
\usage{
Question(text, type = NULL, id = hexaHash(text), answer = NULL,
  hidden.data = quote({ }), data = quote({ }), feedback = "",
  points = 1, dist = 2, epsilon = 1e-04)
}
\arguments{
\item{text}{Body of question}

\item{type}{Type of question}

\item{id}{Id of question}

\item{answer}{Answer of question}
}
\description{
Create a question object
}

