% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizR.R
\name{Quiz}
\alias{Quiz}
\title{Create a quiz object}
\usage{
Quiz(title, groups, data, hidden.data, seed = NULL, hidden.seed = NULL)
}
\arguments{
\item{title}{Title of the quiz}

\item{groups}{List of groups}

\item{data}{Language object defining the data}

\item{hidden.data}{Language object defining the hidden data}

\item{seed}{The seed used to generate data}

\item{hidden.seed}{The seed used to generate hidden data}
}
\value{
A quiz object
}
\description{
Create a quiz object
}

